+++ CHANGELOG.md
@@ -4,8 +4,8 @@
 - Show only time when show percentage is disabled and show time is enabled
 - Made sure progress does not go over 100%
 - Added option to log the changes to hover text
-- Moved settings: "Amount of Decimal Places" and "Show Time" are moved from "2 - General" to "2 - Progress"
-- Attempt to read deprecated settings and remove them (NOT WORKING)
+- Changed settings: "Amount of Decimal Places" and "Show Time" are moved from "2 - General" to "2 - Progress"
+- Read deprecated settings and remove them
 ## v1.5.1
 - Added option to show time remaining
 ## v1.5.0